requirements.txt
# ========================================
# AudioPredictorGenML - Dependencies
# Masters Research Project
# Python 3.9+
# ========================================

# ----------------------------------------
# Deep Learning Framework
# ----------------------------------------
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# ----------------------------------------
# Audio Processing
# ----------------------------------------
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
resampy==0.4.2
pydub==0.25.1

# ----------------------------------------
# Numerical Computing
# ----------------------------------------
numpy==1.24.3
scipy==1.11.2

# ----------------------------------------
# Machine Learning
# ----------------------------------------
scikit-learn==1.3.1

# ----------------------------------------
# Visualization
# ----------------------------------------
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0

# ----------------------------------------
# Data Handling
# ----------------------------------------
pandas==2.1.3
h5py==3.10.0

# ----------------------------------------
# Configuration & Utilities
# ----------------------------------------
pyyaml==6.0.1
python-dotenv==1.0.0

# ----------------------------------------
# Progress & Logging
# ----------------------------------------
tqdm==4.66.1
coloredlogs==15.0.1

# ----------------------------------------
# Development & Interactive
# ----------------------------------------
jupyter==1.0.0
ipython==8.15.0
ipykernel==6.26.0
ipywidgets==8.1.1

# ----------------------------------------
# Testing
# ----------------------------------------
pytest==7.4.3
pytest-cov==4.1.0

# ----------------------------------------
# Code Quality
# ----------------------------------------
black==23.11.0
flake8==6.1.0
pylint==3.0.2
mypy==1.7.1

# ----------------------------------------
# Documentation
# ----------------------------------------
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# ----------------------------------------
# Experiment Tracking (Optional)
# ----------------------------------------
# tensorboard==2.15.1
# wandb==0.16.0

# ----------------------------------------
# Audio Metrics & Evaluation
# ----------------------------------------
pesq==0.0.4
pystoi==0.3.3

# ----------------------------------------
# Additional Utilities
# ----------------------------------------
requests==2.31.0
Pillow==10.1.0
